+++ src/plugins/intel_gna/src/pre_post_process/avx2.hpp
@@ -22,36 +22,36 @@ namespace pre_post_processing {
 
 void convert_matrix_fp32_to_int16_avx(int16_t* ptr_dst,
                                       const float* ptr_src,
-                                      const uint32_t num_rows,
-                                      const uint32_t num_columns,
+                                      const size_t num_rows,
+                                      const size_t num_columns,
                                       const float scale_factor,
                                       bool transpose);
 
 void convert_matrix_fp32_to_int8_avx(int8_t* ptr_dst,
                                      const float* ptr_src,
-                                     const uint32_t num_rows,
-                                     const uint32_t num_columns,
+                                     const size_t num_rows,
+                                     const size_t num_columns,
                                      const float scale_factor,
                                      bool transpose);
 
 void convert_matrix_int32_to_fp32_avx(float* ptr_dst,
                                       const int32_t* ptr_src,
-                                      uint32_t num_rows,
-                                      uint32_t num_columns,
+                                      size_t num_rows,
+                                      size_t num_columns,
                                       float scale_factor,
                                       bool transpose);
 
 void convert_matrix_int16_to_fp32_avx(float* ptr_dst,
                                       const int16_t* ptr_src,
-                                      uint32_t num_rows,
-                                      uint32_t num_columns,
+                                      size_t num_rows,
+                                      size_t num_columns,
                                       float scale_factor,
                                       bool transpose);
 
 void convert_matrix_int8_to_fp32_avx(float* ptr_dst,
                                      const int8_t* ptr_src,
-                                     uint32_t num_rows,
-                                     uint32_t num_columns,
+                                     size_t num_rows,
+                                     size_t num_columns,
                                      float scale_factor,
                                      bool transpose);
 }  // namespace pre_post_processing
